requirements.txt
# Core dependencies
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.2

# AWS
boto3==1.34.10
s3fs==2023.12.2

# Snowflake
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.0
cryptography==41.0.7

# dbt
dbt-core==1.7.7
dbt-snowflake==1.7.1

# Airflow
apache-airflow==2.8.1
apache-airflow-providers-amazon==8.13.0
apache-airflow-providers-snowflake==5.6.0

# Kaggle
kaggle==1.5.16

# Data validation
great-expectations==0.18.8

# Neo4j
neo4j==5.15.0

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Utilities
requests==2.31.0
pyyaml==6.0.1
python-dateutil==2.8.2

# Async web scraping (trio-based)
trio==0.23.2
trio-typing==0.9.0
asks==3.2.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0

